move transducer module into transducers.rs
